xong insert công việc, đang lỗi insert congty

+++ serverInsert_CongTy.php
@@ -1,4 +1,4 @@
-\<?php
+<?php
 /**
  * Created by PhpStorm.
  * User: HieuUit
@@ -6,6 +6,8 @@
  * Time: 11:44 AM
  */
 	
+$MaUser = $_SESSION['mauser'];
+$Username = $_SESSION['username'];
 
 	// initializing variables
 
@@ -14,50 +16,49 @@
 	// connect to the database
 	$db = mysqli_connect('localhost', 'root', '', 'dataweb');
 	
-	
+	mysqli_set_charset($db, "utf8");
 		// REGISTER USER
-		if (isset($_POST['Upload'])) {
+		if (isset($_POST['Insert_CongTy'])) {
 		// receive all input values from the form
 		$TenCty =  $_POST['ten_cty'];
 		$DiaChi = $_POST['dia_chi'];
 		$TGLamViec =  $_POST['tg'];
 		$DaiNgo =  $_POST['dai_ngo'];
-		
+		var_dump($TenCty);
+		var_dump($TGLamViec);
+		var_dump($DiaChi);
+		var_dump($DaiNgo);
+
 		//xu ly logo cong ty
-		$logoname = $_FILES['img_logo']['name'];
-		$logotmp_name = $_FILES['nameimage']['tmp_name'];
+		
+		
+		$logoname = $_FILES['logo_congty']['name'];
+var_dump($logoname);		die();
+		$logotmp_name = $_FILES['logo_congty']['tmp_name'];
 		$folder = 'images/';
 		
 		//xu ly logo cong ty
-		$Ctyname = $_FILES['img_cty']['name'];
-		$Ctytmp_name = $_FILES['nameimage']['tmp_name'];
-		$MaCongTy = (int)$result['MaCongTy'];
+		$Ctyname = $_FILES['anh_congty']['name'];
+		$Ctytmp_name = $_FILES['anh_congty']['tmp_name'];
 		
-		// form validation: ensure that the form is correctly filled ...
-		// by adding (array_push()) corresponding error unto $errors array
-		if (empty($TenCty)) { array_push($errors, "chưa có tên công ty"); }
-		if (empty($DiaChi)) { array_push($errors, "chưa có địa chỉ"); }
-		if (empty($TGLamViec)) { array_push($errors, "chưa có thời gian hoạt động"); }
-		if (empty($DaiNgo)) { array_push($errors, "chưa có đãi ngộ"); }
+	
 		
-		// first check the database to make sure
-		
-		mysqli_set_charset($db, "utf8");
 		// Finally, register user if there are no errors in the form
 		
 		//get the id congty
-		$conn = mysqli_query($db,"select * from user where Username = '$user'");
-		$arrUser = mysqli_fetch_array($conn);
-		//lay ma user
-		$MaUser = (int)$arrUser['MaUser'];
+		if (file_exists($_FILES['logo_congty']['tmp_name'])) {
 		
-		//day tap tin anh len data
-		move_uploaded_file($logotmp_name, $folder.$logoname);
-		move_uploaded_file($Ctytmp_name, $folder.$Ctyname);
+			move_uploaded_file($logotmp_name, $folder.$logoname);
+		}
+		if (file_exists($_FILES['anh_congty']['tmp_name'])) {
+		
+			move_uploaded_file($Ctytmp_name, $folder.$Ctyname);
+		}
+
         $query = "INSERT INTO congty (TenCongTy,DiaChiCongTy, TGLamViecCongTy, DaiNgoCongTy, IconCongTy, AnhCongTy, IDuser) 
   			  VALUES('$TenCty', '$DiaChi', '$TGLamViec','$DaiNgo', '/images/$logoname', '/images/$Ctyname' '$MaUser')";
 		
-
+			var_dump($query);die();
 			mysqli_query($db, $query);
 
         ?>
